+++ core/src/main/java/org/jdbi/v3/core/Query.java
@@ -23,7 +23,7 @@ import java.util.Locale;
 
 import org.jdbi.v3.core.exception.ResultSetException;
 import org.jdbi.v3.core.exception.UnableToExecuteStatementException;
-import org.jdbi.v3.core.mapper.BeanMapper;
+import org.jdbi.v3.core.mapper.reflect.BeanMapper;
 import org.jdbi.v3.core.mapper.RowMappers;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.mapper.RowMapperFactory;
