+++ core/src/main/java/org/jdbi/v3/core/unstable/oracle/OracleReturning.java
@@ -21,7 +21,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.jdbi.v3.core.StatementContext;
-import org.jdbi.v3.core.mapper.BeanMapper;
+import org.jdbi.v3.core.mapper.reflect.BeanMapper;
 import org.jdbi.v3.core.exception.ResultSetException;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.statement.StatementCustomizer;
