+++ core/src/test/java/org/jdbi/v3/core/mapper/JoinRowMapperTest.java
@@ -23,6 +23,7 @@ import com.google.common.collect.Multimap;
 
 import org.jdbi.v3.core.H2DatabaseRule;
 import org.jdbi.v3.core.Handle;
+import org.jdbi.v3.core.mapper.reflect.ConstructorMapper;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
